Moved groupByNotNull
